+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.0.0 - 2019-06-26
+## [Unreleased]
+
+### Added
+
+* Client: (annotator) right-mouse button with delete tool sweeps out a rectangle
+  and deletes its contents.
+
+* Client: (annotator) "ctrl-=" cycles through zooms (user, fit-width, fit-height). Also added this shortcut to the key-help list.
+
+### Changed
+
+### Fixed
+
+
+## 0.1.0 - 2019-06-26
 
 This is the first release of PLOM, PaperLessOpenMarking.
+
+
+[Unreleased]: https://gitlab.math.ubc.ca/andrewr/MLP/compare/v0.1.0...master